✏️ Rename: all component file names to upper camel case [#4]
