+++ src/lib/components/header/Icon.svelte
